Finished Post Creation with File Upload -> Problem: API Points -> componentPath: "\src\components\postComponent\post-component.svelte" At the moment in search folder
