+++ src/components/postComponent/components/textarea.svelte
@@ -1,7 +1,10 @@
 <script lang="ts">
-    // Initialize the text variable for storing textarea input
-    let text = '';
+
+    import {
+            postText
+        } from '$stores';    
     
+    var localPostText:string = "";
     // Maximum characters allowed in the textarea
     const maxChars = 256;
 
@@ -14,18 +17,16 @@
     const charsWarning = 50;
 
     // Reactive declarations to update the textarea class based on text length
-    $: classtext = text.length > maxChars - charsWarning && text.length < maxChars
+    $: classtext = localPostText.length > maxChars - charsWarning && localPostText.length < maxChars
         ? classWarning
-        : text.length >= maxChars
+        : localPostText.length >= maxChars
         ? classMaxReached
         : classNormal;
 
-    // Reactive declaration to set a boolean based on whether there's any text
-    $: inputBoolean = text.length > 0;
-
     // Function to handle input changes, update the text, and adjust the textarea height
     function handleInput(event: any) {
-        text = event.target.value;
+        $postText = localPostText;
+        localPostText = event.target.value;
         event.target.style.height = 'inherit';
         event.target.style.height = `${event.target.scrollHeight}px`;
     }
@@ -33,16 +34,15 @@
 
 <div class="container">
 
-<textarea class="textarea {classtext}" rows="3" placeholder="Bitte geben sie hier ihren Post mit maximal 256 Zeichen ein." maxlength={maxChars} bind:value={text} on:input={handleInput}></textarea>
+<textarea class="textarea {classtext}" rows="3" placeholder="Bitte geben sie hier ihren Post mit maximal 256 Zeichen ein." maxlength={maxChars} bind:value={localPostText} on:input={handleInput}></textarea>
 
-
-    {#if text.length === maxChars}
+    {#if localPostText.length === maxChars}
         <!-- Displayed when the character limit is reached -->
         <div class="{classtext}">Du hast das Zeichenlimit erreicht.</div>
     {/if}
 
     <!-- Display the number of remaining characters -->
-    <div class="{classtext}">Verbleibende Zeichen: {maxChars - text.length}</div>
+    <div class="{classtext}">Verbleibende Zeichen: {maxChars - localPostText.length}</div>
 </div>
 
 <style>
@@ -51,7 +51,7 @@
         display: flex;
         flex-direction: column;
         align-items: center;
-        max-width: 500px;
+        max-width: 31.25rem;
         margin: auto;
         margin-bottom: 2rem;
     }
