+++ src/domains/Post.ts
@@ -1,0 +1,4 @@
+export interface UploadRestrictions {
+	type: string;
+	value: string;
+}
