+++ src/domains/index.ts
@@ -11,3 +11,7 @@ export type {
 
 //Server Domains
 export type { Imprint } from './ServerDomains/imprint';
+
+//Upload Restrictions
+export type { UploadRestrictions } from './Post';
+
