+++ src/static/components/index.ts
@@ -1,1 +1,3 @@
 export { personsData, logos } from './about/About.ts';
+export { uploadRestrictions } from './postComponent/Post.ts';
+
