+++ src/static/components/postComponent/Post.ts
@@ -1,0 +1,6 @@
+import type { UploadRestrictions } from '$domains';
+export const uploadRestrictions:UploadRestrictions[] = [
+    { type: "Erlaubte Dateitypen:", value: "jpeg, webp" },
+    { type: "Maximale Dateigröße:", value: "5 MB" },
+    { type: "Maximale Dateianzahl:", value: "1" }
+];
