+++ src/stores/index.ts
@@ -1,2 +1,3 @@
 export { authToken, currentUser, isLoggedIn } from './client';
 export { loading } from './loading';
+export { files,isFileUploaded,uploadedImageUrl,isEditing,isFileSelected,isOverlayVisible,postText } from './post';
